move adding filedescriptor logic to PackageViewModel
