+++ PackageExplorer/PackageViewer.xaml.cs
@@ -8,7 +8,6 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
 using System.Windows.Input;
-using NuGet.Packaging;
 using NuGetPackageExplorer.Types;
 using PackageExplorer.Properties;
 using PackageExplorerViewModel;
@@ -583,39 +582,8 @@ namespace PackageExplorer
             {
                 folder = folder ?? RootFolder;
 
-                foreach (var fileDescription in NativeDragDrop.GetFileGroupDescriptorW(data))
-                {
-                    var parts = fileDescription.Key.Split(Path.DirectorySeparatorChar);
-
-                    var name = parts[parts.Length - 1];
-                    var parentFolder = folder;
-                    for (var i = 0; i < parts.Length - 1; i++)
-                    {
-                        parentFolder = (PackageFolder)parentFolder[parts[i]];
-                    }
-
-                    if (fileDescription.Value != null)
-                    {
-                        var tempFile = Path.Combine(Path.GetTempPath(), name);
-                        using (var stream = fileDescription.Value)
-                        using (var fileStream = File.OpenWrite(tempFile))
-                        {
-                            stream.CopyTo(fileStream);
-                        }
-
-                        var physicalFile = new PhysicalPackageFile
-                        {
-                            SourcePath = tempFile,
-                            TargetPath = name,
-                        };
-
-                        parentFolder.AddFile(new PackageFile(physicalFile, name, parentFolder));
-                    }
-                    else
-                    {
-                        parentFolder.AddFolder(name);
-                    }
-                }
+                var viewModel = DataContext as PackageViewModel;
+                viewModel.AddDraggedAndDroppedFileDescriptors(folder, NativeDragDrop.GetFileGroupDescriptorW(data));
                 return true;
             }
             if (data.GetDataPresent(DataFormats.FileDrop))
